word-to-pdf/backend/requirements.txt
appscript==1.4.0
blinker==1.9.0
click==8.1.8
docx2pdf==0.1.8
Flask==3.1.2
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
tqdm==4.67.1
Werkzeug==3.1.5
zipp==3.23.0
